backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.3.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9

# Storage
minio==7.2.7

# OCR
google-cloud-vision==3.7.2
pdf2image==1.17.0
pypdf==4.2.0

# AI / Embeddings
openai==1.30.1
anthropic==0.28.0
tiktoken==0.7.0

# Background jobs
arq==0.25.0
redis[hiredis]==5.0.4

# Rate limiting
slowapi==0.1.9
limits==3.10.1

# Utilities
python-dotenv==1.0.1
Pillow==10.3.0
aiofiles==23.2.1
httpx==0.27.0
structlog==24.1.0
